inv-planner-backend/requirements.txt
google-generativeai
reportlab
flask
flask-cors
pymongo
python-dotenv
ipython
pycryptodome